+++ src/content/content.ts
@@ -1,5 +1,5 @@
 import '@webcomponents/custom-elements'
-import uuid from './utils/uuid'
+import uuid from '../utils/uuid'
 import PopupTabSwitcher from './popup-tab-switcher'
 
 const existingEl = document.querySelector('#popup-tab-switcher')
