+++ src/content/popup-tab-switcher.scss
