+++ src/content/readme.md
@@ -1,0 +1,4 @@
+## About
+
+Contains the content script part of the extension. It renders popup and listens for user actions to
+notify background script and trigger tab switching.
